backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
flask-jwt-extended==4.5.3
python-dotenv==1.0.0
requests==2.31.0
faiss-cpu==1.13.2
numpy>=1.26.0
python-docx==0.8.11
PyPDF2==3.0.1
sentence-transformers==2.2.2
pyodbc==5.0.1
werkzeug==3.0.1
